requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
backtrader==1.9.75.123
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
gunicorn==20.0.4
idna 
importlib-metadata==1.7.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.2.0
lark-parser==0.9.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib 
mistune==0.8.4
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.7
numpy==1.18.5
packaging==20.4
pandas 
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.5.1
psycopg2==2.8.5
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests 
Send2Trash==1.5.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
Flask-Cors==3.0.8
yfinance==0.1.54


